+++ main.py
@@ -1,5 +1,6 @@
 import numpy as np
 import csv
+from sgd.model import Model
 from sgd.losses import MSE, CE
 from sgd.layers import Linear, AddBias, Sigmoid, ReLU
 from sgd.sgd import sgd
@@ -10,23 +11,10 @@ log = logging.getLogger(__name__)
 LEARNING_RATE = 0.05
 np.random.seed(0)
 
-class Model():
-    def __init__(self, transform=None):
-        self.next = transform
-        return
-
-    def add_next(self, transform):
-        x = self
-        while x.next is not None:
-            x = x.next
-        x.next = transform
-        return
-
 if __name__ == '__main__':
     with open("dataset.csv", "r") as f:
         dataset = np.array(list(csv.reader(f, delimiter=",")), dtype=np.float64)
 
-    #layers = [Layer(i, o, ACTIVATION_DICT[a]) for i, o, a in zip(LAYER_LIST, LAYER_LIST[1:], ACTIVATION_LIST)]
     #loss = MSE()
     loss = CE()
     #dataset[dataset[:, 2] == 0, 2] = -1
@@ -52,20 +40,15 @@ if __name__ == '__main__':
     model.add_next(Linear(2, 15));model.add_next(AddBias(15));model.add_next(ReLU())
     model.add_next(Linear(15, 15));model.add_next(AddBias(15));model.add_next(Sigmoid())
     model.add_next(Linear(15, 15));model.add_next(AddBias(15));model.add_next(ReLU())
-    #model.add_next(Linear(15, 15));model.add_next(AddBias(15));model.add_next(ReLU())
-    #model.add_next(Linear(6, 3));model.add_next(AddBias(3));model.add_next(Sigmoid())
-    #model.add_next(Linear(15, 7));model.add_next(AddBias(7));model.add_next(Sigmoid())
-    #model.add_next(Linear(7, 1));model.add_next(AddBias(1));model.add_next(Sigmoid())
-    #model.add_next(Linear(3, 1));model.add_next(AddBias(1));model.add_next(Sigmoid())
     model.add_next(Linear(15, 1));model.add_next(AddBias(1));model.add_next(Sigmoid())
     sgd(dataset[:, :2], dataset[:, 2:],
     #sgd(np.array([[0.05, 0.10]]),
     #    np.array([[0.01, 0.99]]),
         model,
         loss,
-        batch_size=32,
+        batch_size=4,
     #    batch_size=1,
-        epochs=250,
+        epochs=15,
         shuffle=True,
         lr=LEARNING_RATE)
     # TESTING
@@ -73,11 +56,7 @@ if __name__ == '__main__':
     x = np.arange(0, 1, 1/100)
     y = np.arange(0, 1, 1/100)
     a = np.transpose([np.tile(x, len(y)), np.repeat(y, len(x))])
-    y_test = a.copy().T
-    layer = model.next
-    while layer is not None:
-       y_test = layer.forward(y_test)
-       layer = layer.next
+    y_test = model.forward(a)
     THR = 0.5 
     y_test[y_test >= THR] = 1
     y_test[y_test < THR] = 0
