+++ sgd/model.py
@@ -1,3 +1,20 @@
 class Model():
-    def __init__(self, layer_or_more=None):
-        if isinstance(
+    def __init__(self, transform=None):
+        self.next = transform
+        return
+
+    def add_next(self, transform):
+        x = self
+        while x.next is not None:
+            x = x.next
+        x.next = transform
+        return
+
+    def forward(self, X):
+        X = X.T
+        layer = self.next
+        while layer is not None:
+            X = layer.forward(X)
+            layer = layer.next
+        return X
+
